Module rotor_http::server

HTTP Server implementation

Currently there is only HTTP/1.x implementation. We want to provide HTTP/2.0 and HTTPS

Structs

Head
Request headers

Parser

Response

Stream
A main stream state machine abstaction

Enums

Accept
Socket acceptor State Machine

BodyKind
The expected type of request body, if any.

RecvMode
This type is returned from headers_received handler of either client client or server protocol handler

RequestError
Error type which is passed to bad_request and emit_error_page

Version
Represents a version of the HTTP spec.

Constants

MAX_CHUNK_HEAD
The maximum length of chunk size line in bytes.

MAX_HEADERS_NUM
The maximum allowed number of headers in a request.

MAX_HEADERS_SIZE
The maximum size of request headers in bytes.

Traits

HttpError
A trait which represents an error which can be formatted as HTTP error page

Server
A handler of server-side HTTP

Type Definitions

Fsm
Shortcut type for server state machines.